auto lock wip
