+++ app/src/main/java/org/privacymatters/safespace/AboutActivity.kt
@@ -102,11 +102,13 @@ class AboutActivity : AppCompatActivity() {
 
     override fun onResume() {
         LockTimer.stop()
+        LockTimer.checkLock(this)
         super.onResume()
     }
 
     override fun onPause() {
-        LockTimer.start(this)
+        LockTimer.stop()
+        LockTimer.start()
         super.onPause()
     }
 }
