+++ app/src/main/res/layout/activity_pdfview.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".document.PDFView">
+    tools:context=".document.PDFActivity">
 <!--    android:background="#000000"-->
 
     <androidx.recyclerview.widget.RecyclerView
